+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.2.1 / TBD
 ==================
 - Restrict scope of typeaheads
+- Maintain search query for `search again` link back
 
 0.2.0 / 2018-07-23
 ==================